Fixed incorrect parsing of svelte.
