+++ lib/html/extract-styles.js
@@ -3,10 +3,10 @@
 const htmlparser = require("htmlparser2");
 const buildSyntaxResolver = require("../syntax/build-syntax-resolver");
 const buildTemplateSyntax = require("../template/syntax");
-const SvelteTokenizer = require("./svelte-tokenizer");
+const JsxLikeTokenizer = require("./jsx-like-tokenizer");
 const { cssSafeSyntax } = require("../syntax/syntaxes");
 
-function iterateCode(source, { onStyleTag, onStyleAttribute, svelte }) {
+function iterateCode(source, { onStyleTag, onStyleAttribute, jsxLike }) {
 	const openTag = {};
 	let disable, ignore, style;
 
@@ -93,7 +93,7 @@ function iterateCode(source, { onStyleTag, onStyleAttribute, svelte }) {
 			},
 		},
 		{
-			Tokenizer: svelte ? SvelteTokenizer : undefined,
+			Tokenizer: jsxLike ? JsxLikeTokenizer : undefined,
 		}
 	);
 
@@ -160,7 +160,7 @@ function extractStyles(source, opts) {
 	iterateCode(source, {
 		onStyleTag,
 		onStyleAttribute,
-		svelte,
+		jsxLike: svelte,
 	});
 
 	return styles;
