+++ package.json
@@ -57,6 +57,7 @@
   },
   "dependencies": {
     "htmlparser2": "^8.0.0",
+    "js-tokens": "^8.0.0",
     "postcss": "^8.4.0",
     "postcss-safe-parser": "^6.0.0"
   },
