+++ test-fixtures/ast/brace.svelte
@@ -4,3 +4,10 @@
 </script>
 <input class:over-100={value > 100} bind:value={value} style="color: {color}">
 <input class:over-100={  ((value > 100) )  } bind:value={value} style="color: {color}">
+
+
+<style>
+input {
+	color: red;
+}
+</style>
